# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CancerGuard API - Cancer Prevention Platform for Canadians
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export type RiskLevel = (typeof RiskLevel)[keyof typeof RiskLevel];

export const RiskLevel = {
  low: "low",
  moderate: "moderate",
  high: "high",
  very_high: "very_high",
} as const;

export type HealthLogType = (typeof HealthLogType)[keyof typeof HealthLogType];

export const HealthLogType = {
  symptom: "symptom",
  appointment: "appointment",
  screening: "screening",
  medication: "medication",
  lifestyle: "lifestyle",
  note: "note",
} as const;

export interface CancerType {
  id: number;
  name: string;
  description: string;
  /** Incidence rate per 100,000 Canadians */
  canadianIncidenceRate: string;
  commonRiskFactors: string[];
  earlySymptoms: string[];
  /** Recommended age to start screening */
  screeningAge: number;
  icon?: string;
}

export interface PreventionAction {
  id: number;
  title: string;
  description: string;
  frequency: string;
  /** Type of healthcare provider (GP, oncologist, etc.) */
  provider?: string;
  urgent: boolean;
}

export interface PreventionPathway {
  id: number;
  cancerTypeId: number;
  cancerTypeName: string;
  riskLevel: RiskLevel;
  riskPercentageMin?: number;
  riskPercentageMax?: number;
  title: string;
  description: string;
  actions: PreventionAction[];
  screeningFrequency: string;
  lifestyleChanges: string[];
  priority: number;
}

export type CreateRiskAssessmentRequestSex =
  (typeof CreateRiskAssessmentRequestSex)[keyof typeof CreateRiskAssessmentRequestSex];

export const CreateRiskAssessmentRequestSex = {
  male: "male",
  female: "female",
  other: "other",
} as const;

export type CreateRiskAssessmentRequestSmokingStatus =
  (typeof CreateRiskAssessmentRequestSmokingStatus)[keyof typeof CreateRiskAssessmentRequestSmokingStatus];

export const CreateRiskAssessmentRequestSmokingStatus = {
  never: "never",
  former: "former",
  current: "current",
} as const;

export type CreateRiskAssessmentRequestAlcoholConsumption =
  (typeof CreateRiskAssessmentRequestAlcoholConsumption)[keyof typeof CreateRiskAssessmentRequestAlcoholConsumption];

export const CreateRiskAssessmentRequestAlcoholConsumption = {
  none: "none",
  light: "light",
  moderate: "moderate",
  heavy: "heavy",
} as const;

export type CreateRiskAssessmentRequestPhysicalActivity =
  (typeof CreateRiskAssessmentRequestPhysicalActivity)[keyof typeof CreateRiskAssessmentRequestPhysicalActivity];

export const CreateRiskAssessmentRequestPhysicalActivity = {
  sedentary: "sedentary",
  light: "light",
  moderate: "moderate",
  active: "active",
} as const;

export type CreateRiskAssessmentRequestDiet =
  (typeof CreateRiskAssessmentRequestDiet)[keyof typeof CreateRiskAssessmentRequestDiet];

export const CreateRiskAssessmentRequestDiet = {
  poor: "poor",
  average: "average",
  good: "good",
  excellent: "excellent",
} as const;

export type CreateRiskAssessmentRequestSunExposure =
  (typeof CreateRiskAssessmentRequestSunExposure)[keyof typeof CreateRiskAssessmentRequestSunExposure];

export const CreateRiskAssessmentRequestSunExposure = {
  minimal: "minimal",
  moderate: "moderate",
  high: "high",
  very_high: "very_high",
} as const;

export interface CreateRiskAssessmentRequest {
  sessionId: string;
  age: number;
  sex: CreateRiskAssessmentRequestSex;
  province?: string;
  /** Cancer types in family history */
  familyHistory?: string[];
  smokingStatus?: CreateRiskAssessmentRequestSmokingStatus;
  alcoholConsumption?: CreateRiskAssessmentRequestAlcoholConsumption;
  physicalActivity?: CreateRiskAssessmentRequestPhysicalActivity;
  bmi?: number;
  diet?: CreateRiskAssessmentRequestDiet;
  sunExposure?: CreateRiskAssessmentRequestSunExposure;
  hpvVaccinated?: boolean;
  lastScreeningDate?: string;
  existingConditions?: string[];
}

export interface RiskFactor {
  cancerTypeId: number;
  cancerTypeName: string;
  riskScore: number;
  riskLevel: RiskLevel;
  contributingFactors: string[];
}

export interface RiskAssessmentResult {
  id: number;
  sessionId: string;
  createdAt: string;
  /** 0-100 overall risk score */
  overallRiskScore: number;
  riskFactors: RiskFactor[];
  recommendedPathways: PreventionPathway[];
  summary: string;
  nextSteps: string[];
}

/**
 * Optional context such as user's risk profile
 */
export type ChatRequestContext = { [key: string]: unknown };

export interface ChatRequest {
  message: string;
  sessionId: string;
  /** Optional context such as user's risk profile */
  context?: ChatRequestContext;
}

export interface ChatResponse {
  response: string;
  sessionId: string;
  sources?: string[];
}

export interface CreateHealthLogRequest {
  sessionId: string;
  logType: HealthLogType;
  title: string;
  description?: string;
  date: string;
  /** Optional numeric or text value */
  value?: string;
  unit?: string;
  tags?: string[];
}

export interface HealthLog {
  id: number;
  sessionId: string;
  logType: HealthLogType;
  title: string;
  description?: string;
  date: string;
  value?: string;
  unit?: string;
  tags?: string[];
  createdAt: string;
}

export type ListPreventionPathwaysParams = {
  cancerTypeId?: number;
  riskLevel?: RiskLevel;
};

export type ListRiskAssessmentsParams = {
  sessionId?: string;
};

export type ListHealthLogsParams = {
  sessionId?: string;
  logType?: HealthLogType;
};