# lib/api-zod/src/generated/types/chatRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CancerGuard API - Cancer Prevention Platform for Canadians
 * OpenAPI spec version: 0.1.0
 */
import type { ChatRequestContext } from "./chatRequestContext";

export interface ChatRequest {
  message: string;
  sessionId: string;
  /** Optional context such as user's risk profile */
  context?: ChatRequestContext;
}