# lib/api-zod/src/generated/types/createHealthLogRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CancerGuard API - Cancer Prevention Platform for Canadians
 * OpenAPI spec version: 0.1.0
 */
import type { HealthLogType } from "./healthLogType";

export interface CreateHealthLogRequest {
  sessionId: string;
  logType: HealthLogType;
  title: string;
  description?: string;
  date: Date;
  /** Optional numeric or text value */
  value?: string;
  unit?: string;
  tags?: string[];
}