# lib/api-zod/src/generated/types/healthLogType.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CancerGuard API - Cancer Prevention Platform for Canadians
 * OpenAPI spec version: 0.1.0
 */

export type HealthLogType = (typeof HealthLogType)[keyof typeof HealthLogType];

export const HealthLogType = {
  symptom: "symptom",
  appointment: "appointment",
  screening: "screening",
  medication: "medication",
  lifestyle: "lifestyle",
  note: "note",
} as const;